+++ Models/House.cs
@@ -6,7 +6,7 @@ namespace cs_gregslist.Models
     public class House
     {
 
-        public House(int bedrooms, int bathrooms, int levels, string description, int year, int price, string imgUrl)
+        public House(int bedrooms, float bathrooms, int levels, string description, int year, float price, string imgUrl)
         {
             Id = Guid.NewGuid().ToString();
             Bedrooms = bedrooms;
@@ -26,7 +26,7 @@ namespace cs_gregslist.Models
 
         [Required]
         [Range(1, int.MaxValue)]
-        public int Bathrooms { get; set; }
+        public float Bathrooms { get; set; }
 
         [Required]
         [Range(1, int.MaxValue)]
@@ -40,7 +40,7 @@ namespace cs_gregslist.Models
 
         [Required]
         [Range(0, int.MaxValue)]
-        public int Price { get; set; }
+        public float Price { get; set; }
 
         public string ImgUrl { get; set; } = "http://www.fillmurray.com/250/250";
